Reset cpo compatible-options in script
We set cpo (compatible-options) to the Vim default value at the
beginning of the script and restore it to the original value at the end
of the script.

This allows us to use the line continuation method (\) in our code, e.g.

iabbrev synchronisation
        \ synchronization

While we are not currently using the line continuation method (\) in our
code, this seems like a good boilerplate piece to have in place (and
appears in the official Vim :help write-plugin)

See :help use-cpo-save
See :help cpo
See :write-plugin