Wrap mapping in hasmapto() check
This check allows a user to define their own :GitHunkToggle mapping in
their .vimrc.

When a customized mapping is set, the default mapping will NOT be
applied.

e.g. of customized mapping in .vimrc

" Map ,h to :GitHunkToggle (in normal mode),
" default mapping in the plugin will not be applied.
nnoremap ,h :GitHunkToggle<CR>

See :help write-plugin (MAPPING)