Add "tags" to .gitignore
During development I'm seeing the file /doc/tags being created.

I don't want to include this file in this project.

I'm not certain why this file is being generated, whether this is
default Vim/Neovim functionality or if my personal configuration is
causing this.

Either way, I don't want the file included in this repo.